# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "all product",
                "description": "Product",
                "operationId": "index_product",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "join"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "/"
                                        },
                                        "organization_id": {
                                            "type": "unsignedBigInteger",
                                            "example": "5"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "lorem"
                                        },
                                        "price": {
                                            "type": "string",
                                            "example": "10000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/product/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "product",
                "description": "Returns project data",
                "operationId": "0d49160165bfa5e700b38835cb61504e",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Post a new data",
                "description": "update  product  data",
                "operationId": "88e67d497d5f16593592b316f2b8dffb",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass category   credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "organization_id",
                                    "description",
                                    "price",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "john"
                                    },
                                    "image": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "bb"
                                    },
                                    "organization_id": {
                                        "type": "number",
                                        "format": "foreign , exists,organizations,id",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "lorem"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "lorem"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "d479af4aa04f4785a2004fba5ae75e21",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "api"
        }
    }
}